[INTERNAL] sap.m: avoid the use of sap.* globals in visual tests
The tests still use globals to share state between tests, but none of
the UI5 global APIs that will be removed in UI5 2.0.

Note: the local scope of the describe function can't be used to store
and share the state as the state lives in the context of the browser
window, not in the context of the test description. These are two
different JavaScript execution contexts.

Change-Id: I198136e25388017cbdd71a3162fa75fa4e85381c