[INTERNAL] MDC: Deprecate legacy attribute support in PropertyHelper with UI5 2.0
In SAPUI5 version 1.121 the "name" property of PropertyInfo was replaced by "key". For compatibility reasons (usage in FE) PropertyInfo definitions using "name" must still work. So the "name" property is also se to the "key" property, if still used.
In UI5 2.0 the "name" property must not loger be supported, so the "name" fallback is behind a @deprecared switch, so it is removed by UI5 2.0 transformation.

In all MDC samples the usage of "name" is replaced by "key".
In the most tests the usage of "name" is replaced by "key". Only some legacy-tests are still using "name" but they are also behind a @deprecate switch to be removed by UI5 2.0 transformation.

Change-Id: Id0b60d43c563b44e1906a83ea2334a0723ab8e12