[FIX] sap.m.MessagePopover: List view is now preserved after resize
MessageView is now keeping its list view after resizing. When a message
popover has only 1 message item, it opens directly in details view.
The issue was that after navigating back to list view and resizing the
popover, the MessageView was not kept in the list view and was
switched back to details view.

SNOW: DINC0648456
Change-Id: I05eab8d66ab137177b6ffb059dcf399826b882e8