[INTERNAL] PluginBase: Parent destruction should deactivate the plugin
 - Detaching the plugin from its parent should deactivate the plugin and
trigger the onDeactivate hook.
 - To support this, PluginBase overrides setParent and expects this API
to be called with null. While this works when the plugin is explicitly
removed, it does not happen when the plugin's parent or descendant is
destroyed.
 - According to the comment in ManagedObject#destroyAggregation, it
seems the core team cannot address this scenario due to compatibility
reasons. To handle it on our side, PluginBase now also overrides the
exit hook and deactivates the plugin.
 - Because the exit hook did not exist previously, some plugins that
override exit were not calling PluginBase#exit. This change ensures that
all plugins inherits from PluginBase and using exit correctly call the
base implementation.
 - Plugins have also been refactored accordingly, since the onDeactivate
hook is now always part of the destroy process.

SNOW: DINC0628789

Change-Id: I7a2b7693cb171c56679b8863434e0db69559436c