[INTERNAL] sap.m.TableTitle: Display rowCount and selection info plus title
This control will enhance the title behavior of the table.
It takes a sap.m.Title as aggregation and displays it as a usual table
title.
In addition the TableTitle can display the current rowCount and the
number
of selected rows of the table. There are two styles which can be used
to change how the rowCount and selectedCount are displayed.

JIRA: CPOUIFTEAMB-2566
Change-Id: I84ffeee6fa504c54f7fca2437b5e701a8833befe