[INTERNAL] sap.ui.fl: Implement Message Broker for Support API
The support tools using SupportAPI are currently not able to handle
scenarios where the application is running in a different context (e.g.
iframe in cFLP). To enable communication between these contexts,
the FLP Message Broker needs to be used.

This change initializes the SupportAPI Message Broker communication
when the application is started in debug mode. This allows support tools
to communicate with the application even when it is running in a different
context.

A related change in the Flex Support Tools repository will follow to
establish the connection from the Flex Support Tools side.

Only the Flex Scripts module supports the Message Broker at the
moment. Other modules like dependencies or overlay info need to be
enabled in follow-up changes.

JIRA: TEAM42-499
Change-Id: I644f4e039e560d3da0a93720739e8a4504c37e60